pyproject.toml
[tool.poetry]
name = "docker-harbormaster"
version = "0.1.18"
description = "A supervisor for docker-compose apps."
authors = ["Stavros Korokithakis <hi@stavros.io>"]
repository = "https://gitlab.com/stavros/harbormaster"
homepage = "https://gitlab.com/stavros/harbormaster"
license = "AGPL-3.0-or-later"
readme = "README.md"

[tool.poetry.scripts]
harbormaster = "docker_harbormaster.cli:cli"

[tool.poetry.dependencies]
python = "^3.6"
PyYAML = "^5.4.1"
click = "^7.1.2"
docker-compose = "^1.29.1"
GitPython = "^3.1.18"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
gitchangelog = {url = "https://github.com/sarnold/gitchangelog/archive/refs/heads/master.zip"}
pystache = {url = "https://github.com/sarnold/pystache/archive/refs/heads/master.zip"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
